Update package.json
Test how flexible ?
Is this only a really polish example posted as a proof of concepts ?
Or can we use it like an easy compact template sorts of nice module. 
And if is it ? If any body can take example on for bootstrap with a short
period of tweaks, changes, iterations and little time of adaptation.
Be an easily comprehensible, and readable pack of code and assets nice to learn.